Changed file structure, moved navigation CSS to separate file
