+++ src/App.css
@@ -31,21 +31,3 @@
     transform: rotate(360deg);
   }
 }
-
-ul.menuHeader li {
-  display: inline;
-  list-style-type: none;
-  margin: 0;
-}
-ul.menuHeader {
-  background-color: #282c34;
-  padding: 0;
-  margin-top: 0;
-}
-ul.menuHeader li a {
-  color: #FFF;
-  font-weight: bold;
-  text-decoration: none;
-  padding: 20px;
-  display: inline-block;
-}
