+++ src/components/common/Navigation.css
@@ -1,0 +1,19 @@
+ul.menuHeader li {
+  display: inline;
+  list-style-type: none;
+  margin: 0;
+}
+
+ul.menuHeader {
+  background-color: #282c34;
+  padding: 0;
+  margin-top: 0;
+}
+
+ul.menuHeader li a {
+  color: #FFF;
+  font-weight: bold;
+  text-decoration: none;
+  padding: 20px;
+  display: inline-block;
+}
