+++ src/components/common/Navigation.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
+import './Navigation.css';
 
 const Navigation = () => {
     return (
