+++ src/components/common/RouteError.js
@@ -1,5 +1,5 @@
 import React from "react";
-import logo from '../logo.svg';
+import logo from '../../logo.svg';
 
 const RouteError = () => {
   return (
