+++ src/components/customers/Customers.js
@@ -1,17 +1,17 @@
 import React from "react";
-import logo from '../logo.svg';
+import logo from '../../logo.svg';
 
-const About = () => {
+const Customers = () => {
   return (
     <div>
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
-          <strong>About</strong>
+          <strong>Customers</strong>
         </p>
       </header>
     </div>
   );
 };
 
-export default About;
+export default Customers;
